Tournament Files renamed for clarity
