+++ client/src/App.js
@@ -7,9 +7,9 @@ import { Provider } from 'react-redux';
 import store from './store';
 import NavigationBar from './components/layout/NavigationBar';
 import ProfileSidebar from './components/layout/ProfileSidebar';
-import TournamentIndex from './components/tournaments/Index';
-import TournamentShow from './components/tournaments/Show';
-import TournamentStart from './components/tournaments/Start';
+import TournamentIndex from './components/tournaments/TournamentIndex';
+import TournamentShow from './components/tournaments/TournamentShow';
+import TournamentStart from './components/tournaments/TournamentStart';
 import PlayerProfile from './components/players/PlayerProfile';
 import PlayerDirectory from './components/players/PlayerDirectory';
 import SmashNews from './components/news/SmashNews';
